+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- include_rules support
+- to make room for this, `:rules` and `:files` keyword arguments are deprecated.
+  Please use `:ignore_rules` and `:ignore_files` instead.
+
 # v0.3.3
 - some performance improvements. maybe